// Front/src/app/funcionarios/model/funcionario.ts
export class Funcionario {
  id: number;
  nome: string;
  ndi: string;
  user_id: number;
  genero: string;
  estado_civil: string;
  tipo_identificacao: string;
  data_nascimento: any;
  base_municipio_id: number;
  endereco: string;
  data_emissao: any;
  banco: string;
  mae: string;
  pai: string;
  telefone_principal: string;
  telefone_alternativo: string;
  outro_contacto: number;

  // base_pessoa_id: number;
  rh_categoria_id: number;
}